apps/api_django/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.1
dj-database-url==2.3.0
python-dotenv==1.0.1
gunicorn==22.0.0




